Add json file for alternate UI & cleanup (#5664)
* small cleanup and add json for William

* More cleanup, add another contractor to employee filter

* Add instructions on how to filter people and add milestone labels

* Comment out writing json file until William ready..
